requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1

